add support for kotlin multiplatform this commit adds support for most kotlin multiplatform targets including ios, wasm, js, jvm also fixes https://github.com/nanihadesuka/LazyColumnScrollbar/issues/41
all tests have passed
compiled for all platform and got successful

* setup targets
* move tests from androidTest to androidUnitTest sourceSet, some dependencies are android-specific and we cant put it on commonTest sourceSet
* moved previews to androidMain sourceSet as it is android-specific
* migrated some components to material3
* bumped Kotlin to 2.0.20
* bumped agp to 8.6.0
* remove explicit compose buildFeature declaration as is is no more needed after AGP 8.5.2
* bumped some dependencies to their latest version
* get rid on some unused dependencies like appcompat etc..
* use jetbrains compose compiler and jetbrains compose dependencies
