+++ .gitignore
@@ -2,6 +2,7 @@
 # Gradle files
 .gradle/
 build/
+.kotlin
 
 # Local configuration file (sdk path, etc)
 local.properties
