+++ app/src/main/java/my/nanihadesuka/lazycolumnscrollbar/ui/theme/Shape.kt
@@ -1,7 +1,7 @@
 package my.nanihadesuka.lazycolumnscrollbar.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
+import androidx.compose.material3.Shapes
 import androidx.compose.ui.unit.dp
 
 val Shapes = Shapes(
