+++ app/src/main/java/my/nanihadesuka/lazycolumnscrollbar/ui/theme/Theme.kt
@@ -1,20 +1,18 @@
 package my.nanihadesuka.lazycolumnscrollbar.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.darkColors
-import androidx.compose.material.lightColors
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 
-private val DarkColorPalette = darkColors(
+private val DarkColorPalette = darkColorScheme(
 	primary = Purple200,
-	primaryVariant = Purple700,
 	secondary = Teal200
 )
 
-private val LightColorPalette = lightColors(
+private val LightColorPalette = lightColorScheme(
 	primary = Purple500,
-	primaryVariant = Purple700,
 	secondary = Teal200
 	
 	/* Other default colors to override
@@ -40,7 +38,7 @@ fun LazyColumnScrollbarTheme(darkTheme: Boolean = isSystemInDarkTheme(), content
 	}
 	
 	MaterialTheme(
-		colors = colors,
+		colorScheme = colors,
 		typography = Typography,
 		shapes = Shapes,
 		content = content
