+++ app/src/main/java/my/nanihadesuka/lazycolumnscrollbar/ui/theme/Type.kt
@@ -1,6 +1,6 @@
 package my.nanihadesuka.lazycolumnscrollbar.ui.theme
 
-import androidx.compose.material.Typography
+import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
@@ -8,7 +8,7 @@ import androidx.compose.ui.unit.sp
 
 // Set of Material typography styles to start with
 val Typography = Typography(
-	body1 = TextStyle(
+	bodyLarge = TextStyle(
 		fontFamily = FontFamily.Default,
 		fontWeight = FontWeight.Normal,
 		fontSize = 16.sp
