+++ build.gradle.kts
@@ -7,7 +7,9 @@ buildscript {
 }
 
 plugins {
-    alias(libs.plugins.kotlin.android) apply false
+    alias(libs.plugins.kotlin.multiplatform) apply false
+    alias(libs.plugins.jetbrains.compose.compiler) apply false
+    alias(libs.plugins.jetbrains.compose) apply false
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.android.library) apply false
 }
@@ -24,8 +26,4 @@ subprojects {
     tasks.withType<Test> {
         maxParallelForks = (Runtime.getRuntime().availableProcessors() - 1).coerceAtLeast(1)
     }
-}
-
-tasks.register("clean", Delete::class) {
-    delete(rootProject.buildDir)
 }
