+++ gradle.properties
@@ -20,3 +20,9 @@ android.enableJetifier=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
 org.gradle.unsafe.configuration-cache=false
+
+kotlin.mpp.androidGradlePluginCompatibility.nowarn=true
+kotlin.native.ignoreDisabledTargets=true
+
+org.jetbrains.compose.experimental.jscanvas.enabled=true
+org.jetbrains.compose.experimental.macos.enabled=true
