+++ lib/src/androidMain/AndroidManifest.xml
