+++ lib/src/androidUnitTest/kotlin/my/nanihadesuka/compose/RowScrollbarTest.kt
@@ -1,12 +1,13 @@
 package my.nanihadesuka.compose
 
-import androidx.compose.foundation.layout.fillMaxSize
+
+import androidx.compose.foundation.ScrollState
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -21,18 +22,17 @@ import org.junit.runner.RunWith
 import org.robolectric.ParameterizedRobolectricTestRunner
 
 @RunWith(ParameterizedRobolectricTestRunner::class)
-class LazyRowScrollbarTest(private val itemCount: Int) {
+class RowScrollbarTest(private val itemCount: Int) {
 
     companion object {
         @JvmStatic
         @ParameterizedRobolectricTestRunner.Parameters(name = " Items count: {0} ")
-        fun parametrizeListItemCount() =
-            listOf<Array<Int>>(
-                arrayOf(0),
-                arrayOf(1),
-                arrayOf(10),
-                arrayOf(100)
-            )
+        fun parametrizeListItemCount() = listOf<Array<Int>>(
+            arrayOf(0),
+            arrayOf(1),
+            arrayOf(10),
+            arrayOf(100)
+        )
     }
 
     @get:Rule
@@ -97,30 +97,13 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     @Test
-    fun `move thumb to the left - with reverse layout`() {
+    fun `move thumb to the right`() {
         if (itemCount == 0) return
 
-        setContent(reverseLayout = true)
-        scrollbarScreen(composeRule) {
-            moveThumbToLeft(1f)
-            assert {
-                isThumbAtLeft()
-                isItemVisible(itemTag = itemTestTag(itemCount - 1))
-            }
-        }
-    }
-
-    @Test
-    fun `move thumb to the right - with reverse layout`() {
-        if (itemCount == 0) return
-
-        setContent(reverseLayout = true)
+        setContent()
         scrollbarScreen(composeRule) {
             moveThumbToRight()
-            assert {
-                isThumbAtRight()
-                isItemVisible(itemTag = itemTestTag(0))
-            }
+            assert { isThumbAtRight() }
         }
     }
 
@@ -132,27 +115,8 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
         scrollbarScreen(composeRule) {
             moveThumbToRight()
             assert { isThumbAtRight() }
-            moveThumbToLeft(1f)
-            assert {
-                isThumbAtLeft()
-                isItemVisible(itemTag = itemTestTag(0))
-            }
-        }
-    }
-
-    @Test
-    fun `move thumb to the right`() {
-        if (itemCount == 0) return
-
-        setContent()
-        scrollbarScreen(composeRule) {
             moveThumbToLeft()
             assert { isThumbAtLeft() }
-            moveThumbToRight()
-            assert {
-                isThumbAtRight()
-                isItemVisible(itemTag = itemTestTag(itemCount - 1))
-            }
         }
     }
 
@@ -162,7 +126,7 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
 
         setContent(indicatorContent = { value, _ -> IndicatorContent(value) })
         scrollbarScreen(composeRule) {
-            moveThumbToRight(startFrom = 0.05f)
+            moveThumbToRight()
             assert { isThumbAtRight(indicatorVisible = true) }
         }
     }
@@ -175,13 +139,13 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
         scrollbarScreen(composeRule) {
             moveThumbToRight()
             assert { isThumbAtRight(indicatorVisible = true) }
-            moveThumbToLeft(1f)
+            moveThumbToLeft()
             assert { isThumbAtLeft(indicatorVisible = true) }
         }
     }
 
     @Test
-    fun `always show scrollbar false`() {
+    fun `always show thumb false`() {
         if (itemCount == 0) return
 
         setContent(
@@ -195,7 +159,7 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     @Test
-    fun `always show scrollbar true`() {
+    fun `always show thumb true`() {
         if (itemCount == 0) return
 
         setContent(
@@ -209,7 +173,7 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     @Test
-    fun `scroll list to the right`() {
+    fun `scroll list to the end`() {
         if (itemCount == 0) return
 
         setContent()
@@ -220,7 +184,7 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     @Test
-    fun `scroll list to the left`() {
+    fun `scroll list to the start`() {
         if (itemCount == 0) return
 
         setContent()
@@ -228,30 +192,6 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
             scrollListToItem(testTag = itemTestTag(itemCount - 1))
             assert { isThumbAtRight() }
             scrollListToItem(testTag = itemTestTag(0))
-            assert { isThumbAtLeft() }
-        }
-    }
-
-    @Test
-    fun `scroll list to the right - with reverse layout`() {
-        if (itemCount == 0) return
-
-        setContent(reverseLayout = true)
-        scrollbarScreen(composeRule) {
-            scrollListToItem(testTag = itemTestTag(itemCount - 1))
-            assert { isThumbAtLeft() }
-            scrollListToItem(testTag = itemTestTag(0))
-            assert { isThumbAtRight() }
-        }
-    }
-
-    @Test
-    fun `scroll list to the left - with reverse layout `() {
-        if (itemCount == 0) return
-
-        setContent(reverseLayout = true)
-        scrollbarScreen(composeRule) {
-            scrollListToItem(testTag = itemTestTag(itemCount - 1))
             assert { isThumbAtLeft() }
         }
     }
@@ -421,7 +361,7 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     @Composable
-    private fun IndicatorContent(value: Int) {
+    private fun IndicatorContent(value: Float) {
         Surface {
             Text(
                 text = "Indicator",
@@ -431,25 +371,24 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
     }
 
     private fun setContent(
-        state: LazyListState = LazyListState(),
+        state: ScrollState = ScrollState(initial = 0),
         side: ScrollbarLayoutSide = ScrollbarLayoutSide.End,
         alwaysShowScrollbar: Boolean = true,
         thickness: Dp = 6.dp,
         padding: Dp = 8.dp,
         thumbMinLength: Float = 0.1f,
-        thumbMaxLength: Float = 1f,
+        thumbMaxLength: Float = 1.0f,
         thumbColor: Color = Color(0xFF2A59B6),
         thumbSelectedColor: Color = Color(0xFF5281CA),
         thumbShape: Shape = CircleShape,
         enabled: Boolean = true,
         selectionMode: ScrollbarSelectionMode = ScrollbarSelectionMode.Thumb,
         selectionActionable: ScrollbarSelectionActionable = ScrollbarSelectionActionable.Always,
-        indicatorContent: (@Composable (index: Int, isThumbSelected: Boolean) -> Unit)? = null,
-        listItemsCount: Int = itemCount,
-        reverseLayout: Boolean = false
+        indicatorContent: (@Composable (normalizedOffset: Float, isThumbSelected: Boolean) -> Unit)? = null,
+        listItemsCount: Int = itemCount
     ) {
         composeRule.setContent {
-            LazyRowScrollbar(
+            RowScrollbar(
                 state = state,
                 settings = ScrollbarSettings(
                     enabled = enabled,
@@ -467,12 +406,8 @@ class LazyRowScrollbarTest(private val itemCount: Int) {
                 ),
                 indicatorContent = indicatorContent,
             ) {
-                LazyRow(
-                    state = state,
-                    modifier = Modifier.fillMaxSize(),
-                    reverseLayout = reverseLayout
-                ) {
-                    items(listItemsCount, key = { it }) {
+                Row(Modifier.verticalScroll(state = state)) {
+                    repeat(listItemsCount) {
                         Text(
                             text = "Item $it",
                             modifier = Modifier
